slackbot/requirements.txt
beautifulsoup4==4.8.1
celery==5.0.5
Django==3.1.7
requests==2.22.0
redis==3.5.3
slackclient==2.9.3


